+++ producer.h
@@ -13,7 +13,7 @@ public:
 
 //    void run();
     QMutex* QMutex_ptr = NULL;
-    QList<int>* QList_ptr = NULL;
+    QList<unsigned long long>* QList_ptr = NULL;
     Producer* producer_ptr = NULL;
     void fill(int times);
